requirements.txt
aiofiles==23.2.1
aiogram==3.7.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
attrs==23.2.0
Brotli==1.1.0
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
frozenlist==1.4.1
gallery_dl==1.27.0
idna==3.7
magic-filter==1.0.12
multidict==6.0.5
mutagen==1.47.0
pillow==10.3.0
pyaes==1.6.1
pycryptodomex==3.20.0
pydantic==2.7.3
pydantic_core==2.18.4
Pyrogram==2.0.106
PySocks==1.7.1
requests==2.32.3
TgCrypto==1.2.5
typing_extensions==4.12.2
urllib3==2.2.1
websockets==12.0
yarl==1.9.4
yt-dlp==2024.5.27
python-dotenv
